requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiogram==2.13.0
    # via -r requirements.in
aiohttp==3.8.4
    # via aiogram
aioredis==2.0.1
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via
    #   aiohttp
    #   aioredis
attrs==23.1.0
    # via aiohttp
babel==2.12.1
    # via aiogram
build==0.10.0
    # via pip-tools
certifi==2023.5.7
    # via aiogram
charset-normalizer==3.1.0
    # via aiohttp
click==8.1.3
    # via
    #   pip-compile-multi
    #   pip-tools
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
hiredis==2.2.3
    # via -r requirements.in
idna==3.4
    # via yarl
loguru==0.7.0
    # via -r requirements.in
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
packaging==23.1
    # via build
pip-compile-multi==2.6.3
    # via -r requirements.in
pip-tools==6.13.0
    # via pip-compile-multi
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via -r requirements.in
tomli==2.0.1
    # via build
toposort==1.10
    # via pip-compile-multi
typing-extensions==4.6.3
    # via aioredis
wheel==0.40.0
    # via pip-tools
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
